调试通过购进新增、提交功能。

+++ mf-item/mf-item-service/src/main/java/com/mallfe/item/controller/ItemController.java
@@ -73,7 +73,7 @@ public class ItemController {
      * @return 商品完整信息
      */
     @GetMapping("queryid")
-    public ResponseEntity<Sp> queryById(@RequestParam("hh") Long id){
+    public ResponseEntity<Sp> queryById(@RequestParam("hh") Integer id){
         Sp s = new Sp();
         s.setId(id);
         return ResponseEntity.ok(itemService.query(s));
